git-svn-id: https://svn.corp.flipkart.com/svn/platform/workflow/trunk@1393 6da837af-d5f2-45fc-8808-b2ac0d525091

+++ src/main/java/flipkart/platform/workflow/node/workstation/WorkStation.java
@@ -37,10 +37,10 @@ abstract class WorkStation<I, O, J extends Job<I>> implements Node<I, O>
     private final String name;
     private final JobFactory<? extends J> jobFactory;
 
-    private final byte maxAttempts;
+    private final int maxAttempts;
 
     public WorkStation(final String name, int numThreads,
-            final byte maxAttempts, final JobFactory<? extends J> jobFactory)
+            final int maxAttempts, final JobFactory<? extends J> jobFactory)
     {
         this.name = name;
         this.jobFactory = jobFactory;
